restructure how tests work
